requirements.txt.txt
fastapi
uvicorn
pytesseract
pillow
httpx
python-multipart
transformers
torch
